changement dans ad et delivery

+++ Consommi-tounsi/src/main/java/tn/esprit/spring/entity/Ad.java
@@ -21,7 +21,7 @@ public class Ad {
 	@Id
 	@GeneratedValue (strategy = GenerationType.IDENTITY)
 	private int id_ad;
-	private int cost;
+	private double cost;
 	private int views;
 	
 	@Temporal(TemporalType.DATE)
@@ -45,7 +45,7 @@ public class Ad {
 		super();
 	}
 	
-	public Ad(int id_ad,int cost,int views, Date start_date,Date end_date, String communication_channel,String targeted_population, String ad_name,boolean status ) {
+	public Ad(int id_ad,double cost,int views, Date start_date,Date end_date, String communication_channel,String targeted_population, String ad_name,boolean status ) {
 		super();
 		this.id_ad = id_ad;
 		this.cost = cost;
@@ -64,10 +64,10 @@ public class Ad {
 	public void setIdad(int id_ad) {
 		this.id_ad = id_ad;
 	}
-	public int getCost() {
+	public double getCost() {
 		return cost;
 	}
-	public void setCost(int cost) {
+	public void setCost(double cost) {
 		this.cost = cost;
 	}
 	public int getViews() {
